+++ docs/changelog.rst
@@ -32,6 +32,7 @@ Changelog
 * :feature:`-` Events related to more gitcoin rounds in Arbitrum are now properly decoded as donations.
 * :feature:`-` Support Monerium integration after the v2 contracts upgrade on Ethereum.
 * :bug:`9122` Properly decode 0x swaps with partial refund
+* :bug:`-` Querying zksynclite balances of empty address will now be handled better by rotki.
 
 * :release:`1.36.1 <2024-11-29>`
 * :bug:`-` Coinbase deposits that appeared as withdrawals should now appear as deposits again. Would need to purge and requery coinbase to fix this.